requirements.txt
numpy
pandas
flask
python-dotenv
scikit-learn
seaborn
pickle

#-e . 